requirements.txt
kagglehub
pandas
scikit-learn
matplotlib
seaborn
mlflow
mlflow-sklearn
